requirements.txt
selenium==4.18.1
pytest==8.0.2
pytest-html==4.1.1
webdriver-manager==4.0.1
python-dotenv==1.0.1
Faker==24.2.0
requests==2.31.0
pytest-xdist==3.5.0
pytest-rerunfailures==13.0.0
allure-pytest==2.13.2
py==1.11.0 